Readd search highlights and return to line
I'm not sure why these things were remove but this re-adds search
highlighting along with the ability to return to the line you were
previously on in a file. Also reorged things a bit and commented out the
section for golang (don't remember what that was for but not needed at
the moment).