Change prompt back to longer directory
Obviously doing a lot of thinking with prompt organization, what data to
display, and where. With bash v4 we'd have the PROMPT_DIRTRIM option
which would solve this back and forth but for now we'll leave prompt as
it was before the preceding commit. Yes I could've just removed/reverted
last commit but there were additional changes to comments and script
organization too.