Add dashed date aliases to mimic the underscored ones
The dashed aliases simply point to the underscored aliases for
portability for now. Once I've determined that nothing is calling on the
underscored aliases, I will replace the underscored aliases with the
dashed ones.

I did this because I prefer commands and aliases to be named with dash
separators, not underscores which I reserve for shell variables.

As a side note, I also improved some comments in this commit.